Remove RGBSET from ST7735
The "newer" Adafruit PyBadges use ST7735 displays
that don't seem to support RGBSET the same way.

I haven't been able to determine if a display is
going to support RGBSET with the current hardware
configuration.

This is definitely just a proof of concept, but it
does work on my newer and older PyBadge LCs.